requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-socketio==5.13.0
edge-tts==7.0.2
pydub==0.25.1
python-multipart==0.0.20
SpeechRecognition==3.14.3
aiosqlite==0.21.0
python-dotenv==1.1.1
openai==1.97.0
jinja2==3.1.6
sqlalchemy==2.0.41
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-jose[cryptography]==3.5.0
bcrypt==3.2.2
# psycopg2==2.9.10
psycopg2-binary==2.9.10
asyncpg==0.30.0
boto3==1.39.9
numpy==2.2.6
langdetect==1.0.9
eng_to_ipa==0.0.2
fastapi-mail==1.5.0
google-auth==2.40.3